[Refactor] BoxOfficeRepository 위치 변경
- 구현체와 같은 파일에 있도록 위치 수정 변경 (가독성 고려)
- 기존의 방식은 Clean Architecture의 개념적 구현을 고려해 Domain 영역에 Repository의 Interface 를 노출하려 했지만, 오히려 가독성을 해친다는 점을 고려함...
