+++ BoxOffice/Sources/Data/Repositories/BoxOfficeRepository.swift
@@ -7,6 +7,18 @@
 
 import Foundation
 
+// MARK: - Protocols
+
+protocol BoxOfficeRepositoryProtocol {
+    typealias DailyBoxOfficeCompletion = (Result<DailyBoxOfficeResponse, NetworkError>) -> Void
+    typealias MovieDetailCompletion = (Result<MovieDetailResponse, NetworkError>) -> Void
+    
+    func fetchDailyBoxOffice(endPoint: MovieEndPoint, completion: @escaping DailyBoxOfficeCompletion)
+    func fetchMovieDetail(endPoint: MovieEndPoint, completion: @escaping MovieDetailCompletion)
+}
+
+// MARK: - BoxOfficeRepository
+
 final class BoxOfficeRepository: BoxOfficeRepositoryProtocol {
     
     // MARK: - Properties
